Phi_3_mini/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
accelerate==1.13.0
    # via -r requirements.in
annotated-doc==0.0.4
    # via typer
anyio==4.13.0
    # via httpx
bitsandbytes==0.49.2
    # via -r requirements.in
build==1.4.2
    # via pip-tools
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
click==8.3.2
    # via
    #   pip-tools
    #   typer
colorama==0.4.6
    # via
    #   build
    #   click
    #   tqdm
filelock==3.25.2
    # via
    #   huggingface-hub
    #   torch
fsspec==2026.3.0
    # via
    #   huggingface-hub
    #   torch
h11==0.16.0
    # via httpcore
hf-xet==1.4.3
    # via
    #   -r requirements.in
    #   huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   -r requirements.in
    #   huggingface-hub
huggingface-hub==1.9.1
    # via
    #   accelerate
    #   tokenizers
    #   transformers
idna==3.11
    # via
    #   anyio
    #   httpx
jinja2==3.1.6
    # via torch
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
networkx==3.6.1
    # via torch
numpy==2.4.4
    # via
    #   accelerate
    #   bitsandbytes
    #   torchvision
    #   transformers
packaging==26.0
    # via
    #   accelerate
    #   bitsandbytes
    #   build
    #   huggingface-hub
    #   transformers
    #   wheel
pillow==12.2.0
    # via torchvision
pip-tools==7.5.3
    # via -r requirements.in
psutil==7.2.2
    # via accelerate
pygments==2.20.0
    # via rich
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dotenv==1.2.2
    # via -r requirements.in
pyyaml==6.0.3
    # via
    #   accelerate
    #   huggingface-hub
    #   transformers
regex==2026.4.4
    # via transformers
rich==14.3.3
    # via typer
safetensors==0.7.0
    # via
    #   accelerate
    #   transformers
shellingham==1.5.4
    # via typer
sympy==1.14.0
    # via torch
tokenizers==0.22.2
    # via transformers
torch==2.11.0
    # via
    #   accelerate
    #   bitsandbytes
    #   torchvision
torchaudio==2.11.0
    # via -r requirements.in
torchvision==0.26.0
    # via -r requirements.in
tqdm==4.67.3
    # via
    #   huggingface-hub
    #   transformers
transformers==5.5.0
    # via -r requirements.in
typer==0.24.1
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   transformers
typing-extensions==4.15.0
    # via
    #   anyio
    #   huggingface-hub
    #   torch
wheel==0.46.3
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
